+++ lib/app/config/theme.dart
@@ -7,29 +7,5 @@ const appRedColor = Color(0xFFF80F30);
 const appGreenColor = Color(0xFF0FF830);
 const appYellowColor = Color(0xFFFFD60A);
 
-final lightScheme = ColorScheme.fromSeed(
-  brightness: Brightness.light,
-  dynamicSchemeVariant: .fidelity,
-  seedColor: appPrimaryColor,
-  onSurface: appSlateColor,
-);
-
-final darkScheme = ColorScheme.fromSeed(
-  brightness: Brightness.dark,
-  dynamicSchemeVariant: .fidelity,
-  seedColor: appPrimaryColor,
-  surface: appSlateColor,
-);
-
-final highContrastTheme = ColorScheme.fromSeed(
-  brightness: Brightness.dark,
-  seedColor: appPrimaryColor,
-  surface: Colors.black,
-  onSurface: Colors.white,
-  primary: Colors.white,
-  onPrimary: Colors.black,
-  secondary: Colors.yellowAccent,
-  onSecondary: Colors.black,
-  surfaceContainer: Colors.black,
-  contrastLevel: 1,
-);
+Color? getSurfaceColor(Color appColor, Brightness brightness) =>
+    brightness == .dark && appColor == appPrimaryColor ? appSlateColor : null;