+++ lib/app/models/app_settings.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:storii/app/config/theme.dart';
 import 'package:storii/app/models/user.dart';
+import 'package:storii/shared/helpers/converters.dart';
 
 part 'app_settings.freezed.dart';
 part 'app_settings.g.dart';
@@ -11,6 +13,10 @@ sealed class AppSettings with _$AppSettings {
     @Default(ThemeMode.system) ThemeMode themeMode,
 
     @Default(false) bool useDynamicColor,
+
+    @ColorConverter() @Default(appPrimaryColor) Color appColor,
+
+    @Default(DynamicSchemeVariant.fidelity) DynamicSchemeVariant schemeVariant,
 
     @Default(false) bool usePureBlack,