+++ lib/app/models/app_settings.freezed.dart
@@ -15,7 +15,7 @@ T _$identity<T>(T value) => value;
 /// @nodoc
 mixin _$AppSettings {
 
- ThemeMode get themeMode; bool get useDynamicColor; bool get usePureBlack; UserDomain? get currentUser; Uri? get serverUrl; int get maxLogs; bool get enableHttpLogs; Duration get syncInterval; Duration get syncIntervalMetered;
+ ThemeMode get themeMode; bool get useDynamicColor;@ColorConverter() Color get appColor; DynamicSchemeVariant get schemeVariant; bool get usePureBlack; UserDomain? get currentUser; Uri? get serverUrl; int get maxLogs; bool get enableHttpLogs; Duration get syncInterval; Duration get syncIntervalMetered;
 /// Create a copy of AppSettings
 /// with the given fields replaced by the non-null parameter values.
 @JsonKey(includeFromJson: false, includeToJson: false)
@@ -28,16 +28,16 @@ $AppSettingsCopyWith<AppSettings> get copyWith => _$AppSettingsCopyWithImpl<AppS
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is AppSettings&&(identical(other.themeMode, themeMode) || other.themeMode == themeMode)&&(identical(other.useDynamicColor, useDynamicColor) || other.useDynamicColor == useDynamicColor)&&(identical(other.usePureBlack, usePureBlack) || other.usePureBlack == usePureBlack)&&(identical(other.currentUser, currentUser) || other.currentUser == currentUser)&&(identical(other.serverUrl, serverUrl) || other.serverUrl == serverUrl)&&(identical(other.maxLogs, maxLogs) || other.maxLogs == maxLogs)&&(identical(other.enableHttpLogs, enableHttpLogs) || other.enableHttpLogs == enableHttpLogs)&&(identical(other.syncInterval, syncInterval) || other.syncInterval == syncInterval)&&(identical(other.syncIntervalMetered, syncIntervalMetered) || other.syncIntervalMetered == syncIntervalMetered));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is AppSettings&&(identical(other.themeMode, themeMode) || other.themeMode == themeMode)&&(identical(other.useDynamicColor, useDynamicColor) || other.useDynamicColor == useDynamicColor)&&(identical(other.appColor, appColor) || other.appColor == appColor)&&(identical(other.schemeVariant, schemeVariant) || other.schemeVariant == schemeVariant)&&(identical(other.usePureBlack, usePureBlack) || other.usePureBlack == usePureBlack)&&(identical(other.currentUser, currentUser) || other.currentUser == currentUser)&&(identical(other.serverUrl, serverUrl) || other.serverUrl == serverUrl)&&(identical(other.maxLogs, maxLogs) || other.maxLogs == maxLogs)&&(identical(other.enableHttpLogs, enableHttpLogs) || other.enableHttpLogs == enableHttpLogs)&&(identical(other.syncInterval, syncInterval) || other.syncInterval == syncInterval)&&(identical(other.syncIntervalMetered, syncIntervalMetered) || other.syncIntervalMetered == syncIntervalMetered));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,themeMode,useDynamicColor,usePureBlack,currentUser,serverUrl,maxLogs,enableHttpLogs,syncInterval,syncIntervalMetered);
+int get hashCode => Object.hash(runtimeType,themeMode,useDynamicColor,appColor,schemeVariant,usePureBlack,currentUser,serverUrl,maxLogs,enableHttpLogs,syncInterval,syncIntervalMetered);
 
 @override
 String toString() {
-  return 'AppSettings(themeMode: $themeMode, useDynamicColor: $useDynamicColor, usePureBlack: $usePureBlack, currentUser: $currentUser, serverUrl: $serverUrl, maxLogs: $maxLogs, enableHttpLogs: $enableHttpLogs, syncInterval: $syncInterval, syncIntervalMetered: $syncIntervalMetered)';
+  return 'AppSettings(themeMode: $themeMode, useDynamicColor: $useDynamicColor, appColor: $appColor, schemeVariant: $schemeVariant, usePureBlack: $usePureBlack, currentUser: $currentUser, serverUrl: $serverUrl, maxLogs: $maxLogs, enableHttpLogs: $enableHttpLogs, syncInterval: $syncInterval, syncIntervalMetered: $syncIntervalMetered)';
 }
 
 
@@ -48,7 +48,7 @@ abstract mixin class $AppSettingsCopyWith<$Res>  {
   factory $AppSettingsCopyWith(AppSettings value, $Res Function(AppSettings) _then) = _$AppSettingsCopyWithImpl;
 @useResult
 $Res call({
- ThemeMode themeMode, bool useDynamicColor, bool usePureBlack, UserDomain? currentUser, Uri? serverUrl, int maxLogs, bool enableHttpLogs, Duration syncInterval, Duration syncIntervalMetered
+ ThemeMode themeMode, bool useDynamicColor,@ColorConverter() Color appColor, DynamicSchemeVariant schemeVariant, bool usePureBlack, UserDomain? currentUser, Uri? serverUrl, int maxLogs, bool enableHttpLogs, Duration syncInterval, Duration syncIntervalMetered
 });
 
 
@@ -65,11 +65,13 @@ class _$AppSettingsCopyWithImpl<$Res>
 
 /// Create a copy of AppSettings
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? themeMode = null,Object? useDynamicColor = null,Object? usePureBlack = null,Object? currentUser = freezed,Object? serverUrl = freezed,Object? maxLogs = null,Object? enableHttpLogs = null,Object? syncInterval = null,Object? syncIntervalMetered = null,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? themeMode = null,Object? useDynamicColor = null,Object? appColor = null,Object? schemeVariant = null,Object? usePureBlack = null,Object? currentUser = freezed,Object? serverUrl = freezed,Object? maxLogs = null,Object? enableHttpLogs = null,Object? syncInterval = null,Object? syncIntervalMetered = null,}) {
   return _then(_self.copyWith(
 themeMode: null == themeMode ? _self.themeMode : themeMode // ignore: cast_nullable_to_non_nullable
 as ThemeMode,useDynamicColor: null == useDynamicColor ? _self.useDynamicColor : useDynamicColor // ignore: cast_nullable_to_non_nullable
-as bool,usePureBlack: null == usePureBlack ? _self.usePureBlack : usePureBlack // ignore: cast_nullable_to_non_nullable
+as bool,appColor: null == appColor ? _self.appColor : appColor // ignore: cast_nullable_to_non_nullable
+as Color,schemeVariant: null == schemeVariant ? _self.schemeVariant : schemeVariant // ignore: cast_nullable_to_non_nullable
+as DynamicSchemeVariant,usePureBlack: null == usePureBlack ? _self.usePureBlack : usePureBlack // ignore: cast_nullable_to_non_nullable
 as bool,currentUser: freezed == currentUser ? _self.currentUser : currentUser // ignore: cast_nullable_to_non_nullable
 as UserDomain?,serverUrl: freezed == serverUrl ? _self.serverUrl : serverUrl // ignore: cast_nullable_to_non_nullable
 as Uri?,maxLogs: null == maxLogs ? _self.maxLogs : maxLogs // ignore: cast_nullable_to_non_nullable
@@ -170,10 +172,10 @@ return $default(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( ThemeMode themeMode,  bool useDynamicColor,  bool usePureBlack,  UserDomain? currentUser,  Uri? serverUrl,  int maxLogs,  bool enableHttpLogs,  Duration syncInterval,  Duration syncIntervalMetered)?  $default,{required TResult orElse(),}) {final _that = this;
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( ThemeMode themeMode,  bool useDynamicColor, @ColorConverter()  Color appColor,  DynamicSchemeVariant schemeVariant,  bool usePureBlack,  UserDomain? currentUser,  Uri? serverUrl,  int maxLogs,  bool enableHttpLogs,  Duration syncInterval,  Duration syncIntervalMetered)?  $default,{required TResult orElse(),}) {final _that = this;
 switch (_that) {
 case _AppSettings() when $default != null:
-return $default(_that.themeMode,_that.useDynamicColor,_that.usePureBlack,_that.currentUser,_that.serverUrl,_that.maxLogs,_that.enableHttpLogs,_that.syncInterval,_that.syncIntervalMetered);case _:
+return $default(_that.themeMode,_that.useDynamicColor,_that.appColor,_that.schemeVariant,_that.usePureBlack,_that.currentUser,_that.serverUrl,_that.maxLogs,_that.enableHttpLogs,_that.syncInterval,_that.syncIntervalMetered);case _:
   return orElse();
 
 }
@@ -191,10 +193,10 @@ return $default(_that.themeMode,_that.useDynamicColor,_that.usePureBlack,_that.c
 /// }
 /// ```
 
-@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( ThemeMode themeMode,  bool useDynamicColor,  bool usePureBlack,  UserDomain? currentUser,  Uri? serverUrl,  int maxLogs,  bool enableHttpLogs,  Duration syncInterval,  Duration syncIntervalMetered)  $default,) {final _that = this;
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( ThemeMode themeMode,  bool useDynamicColor, @ColorConverter()  Color appColor,  DynamicSchemeVariant schemeVariant,  bool usePureBlack,  UserDomain? currentUser,  Uri? serverUrl,  int maxLogs,  bool enableHttpLogs,  Duration syncInterval,  Duration syncIntervalMetered)  $default,) {final _that = this;
 switch (_that) {
 case _AppSettings():
-return $default(_that.themeMode,_that.useDynamicColor,_that.usePureBlack,_that.currentUser,_that.serverUrl,_that.maxLogs,_that.enableHttpLogs,_that.syncInterval,_that.syncIntervalMetered);}
+return $default(_that.themeMode,_that.useDynamicColor,_that.appColor,_that.schemeVariant,_that.usePureBlack,_that.currentUser,_that.serverUrl,_that.maxLogs,_that.enableHttpLogs,_that.syncInterval,_that.syncIntervalMetered);}
 }
 /// A variant of `when` that fallback to returning `null`
 ///
@@ -208,10 +210,10 @@ return $default(_that.themeMode,_that.useDynamicColor,_that.usePureBlack,_that.c
 /// }
 /// ```
 
-@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( ThemeMode themeMode,  bool useDynamicColor,  bool usePureBlack,  UserDomain? currentUser,  Uri? serverUrl,  int maxLogs,  bool enableHttpLogs,  Duration syncInterval,  Duration syncIntervalMetered)?  $default,) {final _that = this;
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( ThemeMode themeMode,  bool useDynamicColor, @ColorConverter()  Color appColor,  DynamicSchemeVariant schemeVariant,  bool usePureBlack,  UserDomain? currentUser,  Uri? serverUrl,  int maxLogs,  bool enableHttpLogs,  Duration syncInterval,  Duration syncIntervalMetered)?  $default,) {final _that = this;
 switch (_that) {
 case _AppSettings() when $default != null:
-return $default(_that.themeMode,_that.useDynamicColor,_that.usePureBlack,_that.currentUser,_that.serverUrl,_that.maxLogs,_that.enableHttpLogs,_that.syncInterval,_that.syncIntervalMetered);case _:
+return $default(_that.themeMode,_that.useDynamicColor,_that.appColor,_that.schemeVariant,_that.usePureBlack,_that.currentUser,_that.serverUrl,_that.maxLogs,_that.enableHttpLogs,_that.syncInterval,_that.syncIntervalMetered);case _:
   return null;
 
 }
@@ -223,11 +225,13 @@ return $default(_that.themeMode,_that.useDynamicColor,_that.usePureBlack,_that.c
 @JsonSerializable()
 
 class _AppSettings implements AppSettings {
-  const _AppSettings({this.themeMode = ThemeMode.system, this.useDynamicColor = false, this.usePureBlack = false, this.currentUser, this.serverUrl, this.maxLogs = 100, this.enableHttpLogs = false, this.syncInterval = const Duration(seconds: 20), this.syncIntervalMetered = const Duration(minutes: 1)});
+  const _AppSettings({this.themeMode = ThemeMode.system, this.useDynamicColor = false, @ColorConverter() this.appColor = appPrimaryColor, this.schemeVariant = DynamicSchemeVariant.fidelity, this.usePureBlack = false, this.currentUser, this.serverUrl, this.maxLogs = 100, this.enableHttpLogs = false, this.syncInterval = const Duration(seconds: 20), this.syncIntervalMetered = const Duration(minutes: 1)});
   factory _AppSettings.fromJson(Map<String, dynamic> json) => _$AppSettingsFromJson(json);
 
 @override@JsonKey() final  ThemeMode themeMode;
 @override@JsonKey() final  bool useDynamicColor;
+@override@JsonKey()@ColorConverter() final  Color appColor;
+@override@JsonKey() final  DynamicSchemeVariant schemeVariant;
 @override@JsonKey() final  bool usePureBlack;
 @override final  UserDomain? currentUser;
 @override final  Uri? serverUrl;
@@ -249,16 +253,16 @@ Map<String, dynamic> toJson() {
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _AppSettings&&(identical(other.themeMode, themeMode) || other.themeMode == themeMode)&&(identical(other.useDynamicColor, useDynamicColor) || other.useDynamicColor == useDynamicColor)&&(identical(other.usePureBlack, usePureBlack) || other.usePureBlack == usePureBlack)&&(identical(other.currentUser, currentUser) || other.currentUser == currentUser)&&(identical(other.serverUrl, serverUrl) || other.serverUrl == serverUrl)&&(identical(other.maxLogs, maxLogs) || other.maxLogs == maxLogs)&&(identical(other.enableHttpLogs, enableHttpLogs) || other.enableHttpLogs == enableHttpLogs)&&(identical(other.syncInterval, syncInterval) || other.syncInterval == syncInterval)&&(identical(other.syncIntervalMetered, syncIntervalMetered) || other.syncIntervalMetered == syncIntervalMetered));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _AppSettings&&(identical(other.themeMode, themeMode) || other.themeMode == themeMode)&&(identical(other.useDynamicColor, useDynamicColor) || other.useDynamicColor == useDynamicColor)&&(identical(other.appColor, appColor) || other.appColor == appColor)&&(identical(other.schemeVariant, schemeVariant) || other.schemeVariant == schemeVariant)&&(identical(other.usePureBlack, usePureBlack) || other.usePureBlack == usePureBlack)&&(identical(other.currentUser, currentUser) || other.currentUser == currentUser)&&(identical(other.serverUrl, serverUrl) || other.serverUrl == serverUrl)&&(identical(other.maxLogs, maxLogs) || other.maxLogs == maxLogs)&&(identical(other.enableHttpLogs, enableHttpLogs) || other.enableHttpLogs == enableHttpLogs)&&(identical(other.syncInterval, syncInterval) || other.syncInterval == syncInterval)&&(identical(other.syncIntervalMetered, syncIntervalMetered) || other.syncIntervalMetered == syncIntervalMetered));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,themeMode,useDynamicColor,usePureBlack,currentUser,serverUrl,maxLogs,enableHttpLogs,syncInterval,syncIntervalMetered);
+int get hashCode => Object.hash(runtimeType,themeMode,useDynamicColor,appColor,schemeVariant,usePureBlack,currentUser,serverUrl,maxLogs,enableHttpLogs,syncInterval,syncIntervalMetered);
 
 @override
 String toString() {
-  return 'AppSettings(themeMode: $themeMode, useDynamicColor: $useDynamicColor, usePureBlack: $usePureBlack, currentUser: $currentUser, serverUrl: $serverUrl, maxLogs: $maxLogs, enableHttpLogs: $enableHttpLogs, syncInterval: $syncInterval, syncIntervalMetered: $syncIntervalMetered)';
+  return 'AppSettings(themeMode: $themeMode, useDynamicColor: $useDynamicColor, appColor: $appColor, schemeVariant: $schemeVariant, usePureBlack: $usePureBlack, currentUser: $currentUser, serverUrl: $serverUrl, maxLogs: $maxLogs, enableHttpLogs: $enableHttpLogs, syncInterval: $syncInterval, syncIntervalMetered: $syncIntervalMetered)';
 }
 
 
@@ -269,7 +273,7 @@ abstract mixin class _$AppSettingsCopyWith<$Res> implements $AppSettingsCopyWith
   factory _$AppSettingsCopyWith(_AppSettings value, $Res Function(_AppSettings) _then) = __$AppSettingsCopyWithImpl;
 @override @useResult
 $Res call({
- ThemeMode themeMode, bool useDynamicColor, bool usePureBlack, UserDomain? currentUser, Uri? serverUrl, int maxLogs, bool enableHttpLogs, Duration syncInterval, Duration syncIntervalMetered
+ ThemeMode themeMode, bool useDynamicColor,@ColorConverter() Color appColor, DynamicSchemeVariant schemeVariant, bool usePureBlack, UserDomain? currentUser, Uri? serverUrl, int maxLogs, bool enableHttpLogs, Duration syncInterval, Duration syncIntervalMetered
 });
 
 
@@ -286,11 +290,13 @@ class __$AppSettingsCopyWithImpl<$Res>
 
 /// Create a copy of AppSettings
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? themeMode = null,Object? useDynamicColor = null,Object? usePureBlack = null,Object? currentUser = freezed,Object? serverUrl = freezed,Object? maxLogs = null,Object? enableHttpLogs = null,Object? syncInterval = null,Object? syncIntervalMetered = null,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? themeMode = null,Object? useDynamicColor = null,Object? appColor = null,Object? schemeVariant = null,Object? usePureBlack = null,Object? currentUser = freezed,Object? serverUrl = freezed,Object? maxLogs = null,Object? enableHttpLogs = null,Object? syncInterval = null,Object? syncIntervalMetered = null,}) {
   return _then(_AppSettings(
 themeMode: null == themeMode ? _self.themeMode : themeMode // ignore: cast_nullable_to_non_nullable
 as ThemeMode,useDynamicColor: null == useDynamicColor ? _self.useDynamicColor : useDynamicColor // ignore: cast_nullable_to_non_nullable
-as bool,usePureBlack: null == usePureBlack ? _self.usePureBlack : usePureBlack // ignore: cast_nullable_to_non_nullable
+as bool,appColor: null == appColor ? _self.appColor : appColor // ignore: cast_nullable_to_non_nullable
+as Color,schemeVariant: null == schemeVariant ? _self.schemeVariant : schemeVariant // ignore: cast_nullable_to_non_nullable
+as DynamicSchemeVariant,usePureBlack: null == usePureBlack ? _self.usePureBlack : usePureBlack // ignore: cast_nullable_to_non_nullable
 as bool,currentUser: freezed == currentUser ? _self.currentUser : currentUser // ignore: cast_nullable_to_non_nullable
 as UserDomain?,serverUrl: freezed == serverUrl ? _self.serverUrl : serverUrl // ignore: cast_nullable_to_non_nullable
 as Uri?,maxLogs: null == maxLogs ? _self.maxLogs : maxLogs // ignore: cast_nullable_to_non_nullable